+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 6.0.0-beta.14
 
+- **bug fixes**: fixed datacenter datasource
 - **code enhancements**: added constants and removed duplicated tests to `backupUnit`, `datacenter`, `lan`, `s3_key`, `firewall`, `server`
-- **code enhancements**: for `pcc`, `group`, `user` and `snapshot`:
+- **code enhancements**: for `pcc`, `group`, `user`, `snapshot`, and `volume` :
   - made tests comprehensive
   - optimized test duration by including both match by id and by name in the same test
   - removed duplicated code from import, data_source, resource and tests files
-- **new features**: import for `snapshot`, data_source for `group`, `user`
+- **new features**: import for `snapshot`, data_source for `group`, `user`, `volume`
 
 ## 6.0.0-beta.13
 - **bug fixes**: fixed issue #112 can't attach existing volume to server after recreating server